backend/pyproject.toml
[tool.poetry]
name = "bbs"
version = "0.1.0"
description = "This is a sample project for technical learning of back-end engineers."
authors = ["takenoco82 <reg.takenoco@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
Flask-SQLAlchemy = "^2.4"
PyMySQL = "^0.9.3"
Flask-Migrate = "^2.5"
flask-marshmallow = "^0.10.1"
uwsgi = "^2.0"
uwsgitop = "^0.11.0"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.2"
rope = "^0.14.0"
freezegun = "^0.3.12"
black = {version = "^19.10b0", allow-prereleases = true}
[tool.black]
exclude = '''
/(
    \.eggs |
    \.git |
    \.hg |
    \.mypy_cache |
    \.nox |
    \.tox |
    \.venv |
    _build |
    buck-out |
    build |
    dist |
    tests
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

